requirements.txt
fastapi[all]
uvicorn[standard]
tortoise-orm[asyncmy]
aerich[toml]
pyjwt
bcrypt
cachetools
redis